[VIGNESH M] Added: Linq queries in NotesController to Authorize with JwtToken Payload UserId for all notes api

+++ Repository/Services/FundooNotesRL.cs
@@ -1,4 +1,5 @@
 ﻿using Common.Models;
+using Common.NotesModels;
 using Repository.Context;
 using Repository.Entities;
 using Repository.Interfaces;
@@ -17,7 +18,7 @@ namespace Repository.Services
         {
             this.context = context;
         }
-        public void CreateNotes(NotesModel model)
+        public void CreateNotes(NotesModel model,long jwtUserId)
         {
             try
             {
@@ -30,7 +31,7 @@ namespace Repository.Services
                     Archive = model.Archive,
                     Pin = model.Pin,
                     CreatedAt = model.CreatedAt,
-                    UserId = model.UserId
+                    UserId = jwtUserId
                 };
                 this.context.Add(notes);
                 this.context.SaveChanges();
@@ -41,11 +42,11 @@ namespace Repository.Services
             }
         }
 
-        public IEnumerable<FundooNotes> GetAllNotes()
+        public IEnumerable<FundooNotes> GetAllNotes(long jwtUserId)
         {
             try
             {
-                return this.context.NotesTable.ToList();
+                return this.context.NotesTable.Where(e => e.UserId == jwtUserId);
             }
             catch (Exception)
             {
@@ -53,11 +54,16 @@ namespace Repository.Services
             }
         }
 
-        public FundooNotes GetNotesWithId(long notesId)
+        public FundooNotes GetNotesWithId(long notesId, long jwtUserId)
         {
             try
             {
-                return this.context.NotesTable.FirstOrDefault(i => i.NotesId == notesId);
+                var validUserId = this.context.NotesTable.Where(e => e.UserId == jwtUserId);
+                if(validUserId != null)
+                {
+                    return this.context.NotesTable.FirstOrDefault(i => i.NotesId == notesId);
+                }
+                return null;
             }
             catch (Exception)
             {
@@ -65,20 +71,23 @@ namespace Repository.Services
             }
         }
 
-        public void UpdateNotes(FundooNotes updateNotes, FundooNotes notes)
+        public void UpdateNotes(FundooNotes updateNotes, UpdateNotesModel notes, long jwtUserId)
         {
             try
             {
-                updateNotes.Title = notes.Title;
-                updateNotes.Message = notes.Message;
-                updateNotes.Color = notes.Color;
-                updateNotes.Image = notes.Image;
-                updateNotes.Archive = notes.Archive;
-                updateNotes.Pin = notes.Pin;
-                updateNotes.Trash = notes.Trash;
-                updateNotes.ModifiedAt = notes.ModifiedAt;
-                
-                this.context.SaveChanges();
+                var validUserId = this.context.NotesTable.Where(e => e.UserId == jwtUserId);
+                if (validUserId != null)
+                {
+                    updateNotes.Title = notes.Title;
+                    updateNotes.Message = notes.Message;
+                    updateNotes.Color = notes.Color;
+                    updateNotes.Image = notes.Image;
+                    updateNotes.Archive = notes.Archive;
+                    updateNotes.Pin = notes.Pin;
+                    updateNotes.Trash = notes.Trash;
+                    updateNotes.ModifiedAt = notes.ModifiedAt;
+                    this.context.SaveChanges();
+                } 
             }
             catch (Exception)
             {
@@ -86,12 +95,16 @@ namespace Repository.Services
             }
         }
 
-        public void DeleteNotes(FundooNotes notes)
+        public void DeleteNotes(FundooNotes notes, long jwtUserId)
         {
             try
             {
-                this.context.NotesTable.Remove(notes);
-                this.context.SaveChanges();
+                var validUserId = this.context.NotesTable.Where(e => e.UserId == jwtUserId);
+                if (validUserId != null)
+                {
+                    this.context.NotesTable.Remove(notes);
+                    this.context.SaveChanges();
+                }  
             }
             catch (Exception)
             {
